Move register product collection tester e2e tests from product collection file to its own file
